pyproject.toml
[project]
name = "ta-mcp"
version = "0.1.0"
description = "MCP server for a learning analytics PostgreSQL database"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.20.0",
    "mcp[cli]>=1.4.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
]
